# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}}},"responses":{"200":{"description":"User successfully logged in"},"400":{"description":"Bad Request Body"},"401":{"description":"Invalid Credentials"}}}},"/api/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"400":{"description":"Bad Request Body"}}}},"/api/users/forgot-password":{"post":{"operationId":"UsersController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"URL for Resetting the Password"}}}},"/api/users/reset-password/{id}":{"patch":{"operationId":"UsersController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDTO"}}}},"responses":{"200":{"description":"URL for Resetting the Password"}}}}},"info":{"title":"Simbiotik Technical Task","description":"Simbiotik API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"Token","type":"http"}},"schemas":{"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"phone_no":{"type":"string"},"company_name":{"type":"string"}},"required":["email","password","phone_no","company_name"]},"UserDto":{"type":"object","properties":{}},"ForgotPasswordDto":{"type":"object","properties":{}},"ResetPasswordDTO":{"type":"object","properties":{}}}}}